backapp/backendapp/requirements.txt
API
djangorestframework
mysqlclient
django-cors-headers